Correction dump-text (#2)
* refacto repértoires

* ajout .gitignore

* corrections dump-text

* fichier test

* corrections + ajout cre_tableau script

* ajout_motif espagnol

+++ PROGRAMMES/script_final.sh
@@ -5,7 +5,8 @@
 #-------------------------------------------------------
 
 
-source fonctions_v4.sh
+source ./bin/fonctions.sh
+source ./bin/cree_tableau.sh
 
 #répertoire des fichiers contenant des urls 
 REP_URLS=$1
@@ -14,7 +15,7 @@ REP_URLS=$1
 TABLEAU_HTML=$2"index.html"
 
 
-MOTIFS="منع|حمل|피임|حبوب|politique"
+MOTIFS="منع|حمل|피임|حبوب|anticoncepción"
 
 
 
@@ -40,9 +41,9 @@ REP_DT=../DUMP-TEXT
 REP_CT=../CONTEXTES
 
 #Vidange des répertoires
-rm $REP_PA/*.html >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_PA" || log_failure "Error:"`cat $REP_LOG/command.log`
-rm $REP_DT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_DT" || log_failure "Error:"`cat $REP_LOG/command.log`
-rm $REP_CT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_CT" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_PA)" ] && rm $REP_PA/*.html >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_PA" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_DT)" ] && rm $REP_DT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_DT" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_CT)" ] && rm $REP_CT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_CT" || log_failure "Error:"`cat $REP_LOG/command.log`
 #exit 0
 #-------------------------------------------------------
 # en-tête du fichier html
@@ -79,7 +80,7 @@ do
       
       #le nom complet du fichier de l'url aspirée
       NOM_FIC_PA=""
-      NOM_FIC_PA_UTF8=""
+      NOM_FIC_DT_UTF8=""
       NOM_FIC_DT=""
       NOM_FIC_CT=""
       NOM_FIC_FW=""
@@ -102,19 +103,18 @@ do
             #détection de l'ENCODAGE    
             detection_encodage;
 
+            #
+            dump_text;
+
             # si encodage = UTF8 alors on fait certains traitements
 				# sinon il faudra aussi les faire et probablement d'autres
             gestion_encodage;
-
-            #
-            dump_text;
 
             #
             comptage_motifs;
 
             #
             extraire_contextes;
-
 
             #
             calcul_index;
